Change ReadDouble to ReadFloat

+++ include/IniReader.h
@@ -11,6 +11,6 @@ public:
 
 	int ReadInt(const std::string& key, int default_) const;
 	bool ReadBool(const std::string& key, bool default_) const;
-	double ReadDouble(const std::string& key, double default_) const;
+	float ReadFloat(const std::string& key, float default_) const;
 	std::string ReadString(const std::string& key, const std::string& default_) const;
 };
